start of refactoring
