+++ casual/casual-caller/src/main/java/se/laz/casual/connection/caller/functions/FunctionNoArg.java
@@ -1,0 +1,7 @@
+package se.laz.casual.connection.caller.functions;
+
+@FunctionalInterface
+public interface FunctionNoArg<R>
+{
+    R apply();
+}
