+++ casual/casual-caller/src/main/java/se/laz/casual/connection/caller/functions/FunctionThrowsResourceException.java
@@ -1,0 +1,12 @@
+package se.laz.casual.connection.caller.functions;
+
+import jakarta.resource.ResourceException;
+import se.laz.casual.jca.CasualConnection;
+
+import java.util.UUID;
+
+@FunctionalInterface
+public interface FunctionThrowsResourceException <R>
+{
+    R apply(CasualConnection connection, UUID execution) throws ResourceException;
+}
